requirements.txt
django-phonenumber-field==7.0.2
Django==3.2.15
django-debug-toolbar==3.8.1
django-cleanup==6.0.0
psycopg2-binary==2.9.5
phonenumbers==8.13.4
Pillow==9.4.0
Faker==16.6.1
openpyxl == 3.0.10
python-dateutil~=2.8.2
django-environ==0.9.0
pandas==1.5.3
pdfkit==1.0.0
gunicorn==20.1.0
